Data - st_StatusData CriticalPoint 4byte -> 2byte, st_DropData MinCount 4Byte -> 1Byte, MaxCount 4Byte -> 2Byte
DBStoreProcedure
- ItemTable에 새로운 Item 저장할떄 DataSheetId를 추가로 저장
- GoldTable에 있는 Gold 긁어오는 클래스 추가
- ItemTable에 있는 Item 긁어오는 클래수 추가

GameServInfo
- 캐릭터 기본 정보 전송 하는 메세지 타입 추가

GameServer
- 캐릭터 접속(EnterGame)하면 캐릭터 기본 정보 클라에게 전송하는 작업 추가 ( 지금은 골드, 인벤토리 아이템 전송 함 )
- 캐릭터 생성하면 기본 Gold Table 생성하는 작업 추가

main
- DBThread 정보 출력

+++ GameServer/GameObjectInfo.h
@@ -66,7 +66,7 @@ struct st_StatInfo
 	int32 HP;
 	int32 MaxHP;
 	int32 Attack;
-	int32 CriticalPoint;
+	int16 CriticalPoint;
 	float Speed;
 };
